+++ src/main.js
@@ -9,7 +9,7 @@ import EventEditComponent from "./components/event-edit.js";
 import NoEventsComponent from "./components/no-events.js";
 import {generateEvents} from "./mock/event.js";
 import {generateFilters} from "./mock/filter.js";
-import {RenderPosition, render} from "./utils.js";
+import {RenderPosition, render} from "./utils/render.js";
 
 const TOTAL_EVENTS = 15;
 const events = generateEvents(TOTAL_EVENTS);
