+++ src/utils.js
@@ -59,28 +59,4 @@ const getRandomTime = () => {
 
 };
 
-const createElement = (template) => {
-  const newElement = document.createElement(`div`);
-  newElement.innerHTML = template;
-
-  return newElement.firstChild;
-
-};
-
-const RenderPosition = {
-  AFTERBEGIN: `afterbegin`,
-  BEFOREEND: `beforeend`,
-};
-
-const render = (container, element, place) => {
-  switch (place) {
-    case (RenderPosition.AFTERBEGIN):
-      container.prepend(element);
-      break;
-    case (RenderPosition.BEFOREEND):
-      container.append(element);
-      break;
-  }
-};
-
-export {getDateTimeFormat, getTimeFormat, getRandomInt, getRandomArrayItem, getShuffleArray, getRandomTime, createElement, RenderPosition, render};
+export {getDateTimeFormat, getTimeFormat, getRandomInt, getRandomArrayItem, getShuffleArray, getRandomTime};
