+++ src/utils/render.js
@@ -1,0 +1,23 @@
+export const createElement = (template) => {
+  const newElement = document.createElement(`div`);
+  newElement.innerHTML = template;
+
+  return newElement.firstChild;
+
+};
+
+export const RenderPosition = {
+  AFTERBEGIN: `afterbegin`,
+  BEFOREEND: `beforeend`,
+};
+
+export const render = (container, element, place) => {
+  switch (place) {
+    case (RenderPosition.AFTERBEGIN):
+      container.prepend(element);
+      break;
+    case (RenderPosition.BEFOREEND):
+      container.append(element);
+      break;
+  }
+};
